% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{sw.usa.bounding.box}
\alias{sw.usa.bounding.box}
\title{Region of interest bounding box}
\format{An object of class \code{list} of length 4.}
\usage{
sw.usa.bounding.box
}
\description{
List with coordinates of bounding box for the southwestern US.
}
\examples{

}
\keyword{datasets}

